Fix `option:set` command

+++ src/Console/Commands/OptionSetCommand.php
@@ -40,7 +40,7 @@ class OptionSetCommand extends Command
     {
         $key = $this->argument('key');
         $value = implode($this->argument('value'), ' ');
-        $value = is_numeric($value) ? (int) $value : $value;
+        $value = is_numeric($value) ? (float) $value : $value;
 
         Option::set($key, $value);
         return 0;
